working to get the save done correctly

+++ signal_utils/common/binary_file_ops.py
@@ -29,17 +29,18 @@ def write_binary_image(filename: Path, data):
         #file_id.write(struct.pack('<I', data))
         np.array(data, dtype=np.uint32).tofile(file_id)
 
-def interleave_iq(data: NDArray[np.complex64]) -> NDArray[np.int16]:
+def interleave_iq(data: NDArray[np.complex128]) -> NDArray[np.int16]:
     data_flat = np.empty(2 * len(data))
     data_flat[0::2] = np.real(data)
     data_flat[1::2] = np.imag(data)
 
     return np.array(data_flat, dtype=np.int16)
 
-def get_iq_bytes(data: NDArray[np.complex64]) -> bytes:
+def get_iq_bytes(data: NDArray[np.complex128]) -> bytes:
     data_flat = interleave_iq(data)
 
     return data_flat.tobytes()
+    # return data_flat
 
 def write_binary_iq_data(filename: Path, data: NDArray[np.complex64]) -> None:
     data_flat = interleave_iq(data)
